Update patch set 1
Patch Set 1: Code-Review-1

This is not ready for primetime. It removes support for tfa9887L and bypasses a couple checks to get the amplifier to initialize. There are still a bunch of bugs to work out and we need to sort out why some of the ready checks fail.

Patch-set: 1