Update patch set 3
Patch Set 3: Code-Review-1

Requires:
https://github.com/CM10DNA/proprietary_vendor_htc/commit/4532d21363ff5cb81b6ad169ccf7d9165b63ad24

It seems to be initializing okay now. 

As for issues: I get some random clicks and pops when starting and stopping audio streams and I had to bypass a couple ready checks and just put waits in there as it wasn't returning the expected values. Not sure if we are supposed to return a different value because our config is different or if at that point the chip is in an unusual state.

As far as I can tell though it does appear to be functioning outside of bypass mode.

Patch-set: 3
Label: Code-Review=-1