Update patch set 3
Patch Set 3:

It also doesn't seem to initialize completely when first booted. It needs to switch DSP modes once or twice (e.g. make a call and turn speakerphone on and then off a couple times then hang up). I believe the m7 has the same issue though and that it's being worked on.

Patch-set: 3