Update patch set 1
Patch Set 1: Code-Review-1

Yea, a little cleanup here

We should determine what gps.conf we want to use.

Also, I believe I only enabled gps/gps-nomsb.conf on jfltevzw so we may or may not want that (if not, it can be deleted). That fixes an issue we had where we couldn't get a lock when 'Wifi and Mobile Network Location" was turned on.

Patch-set: 1