Update patch set 1
Patch Set 1:

So I cherry-picked this commit and built it and tried all three gps.conf files and it doesn't seem to be working at all for me (using GPStatus I see 0/0 satellites).

Are there some vendor blobs that I am missing or anything?  I'm not seeing anything useful in the logs.

Patch-set: 1