Update patch set 2
Patch Set 2:

I am building this now just to make sure it works well for me, but I'm sure it will so no need to wait for that.

Although, I don't have any easy way to test the lock either.

Didn't someone say that they were getting a replacement phone in the next day or two and would be willing to test this?  If so, maybe we should just use that offer?

Patch-set: 2