Update patch set 2
Patch Set 2:

This is what I was looking at last night so...

time spent debugging: 1

but only "1".  I didn't have much time to spend on it.

I was going down the path that there was some issue talking with the wpa_supplicant process.  logcat shows that it fails to store the country code which I was guessing stops it from doing anything because it hasn't decided which bands it is permitted to use.

I added ALOGE() messages to all the error paths for sending a message (in hardware/libhardware_legacy/wifi/wifi.c).  I didn't see any error messages when using this code.

I have no idea if that means I am in the wrong file, if the logging doesn't work in that file or if I'm completely off on where the problem lies.

Patch-set: 2