Update patch set 1
Patch Set 1: Code-Review+1

Dan,

This chain of commits is my attempt to port Matt Mower's cleanup of the M7 snd_soc_msm config to our device tree.  In doing so, I found a few things that were different and that shouldn't be so I also added commits to fix them.

Patch-set: 1
Label: Code-Review=+1