Update patch set 1
Patch Set 1:

I ran this for a couple of hours last night with nothing negative to report.

One positive change from this is that the grey artifact when exiting a netflix video (in portrait orientation) is fixed by switching to these libraries.  It was present with both the old blobs and the Moto X blobs.

Patch-set: 1
Label: Code-Review=0