Update patch set 1
Patch Set 1:

I don't actually have a USB dock to test this with.  All I have is a wired USB headset which includes a USB adapter.  Before this commit, it wouldn't play audio over this headset.  After this commit it plays it but only at 0 or 100% volume which is how it worked in CM 10.2.

Patch-set: 1