dlx: tfa9887: Attempting to get the amp to startup correctly
This commit has some hacks in it to ignore error cases (such as
the critically important pll/clks lock).

It does, however, cause the same i2c register writes on boot
as the 2.06 stock ROM.  That is, if you instrument the ic2
register writes, everything up to the pll lock is identical.

On the stock ROM, however, the boot animation plays an
audio file.  I suspect that this is causing some other
initialization of something.  I'm not sure how to replicate
it.

NOTE: We should not be sending the dsp bypass here.  I was
trying that to hack things up to work and left it here as
a reminder that it does not work.

NOTE: We probably need our own vendor blobs and cannot reuse
the ones from the M7.  The M7 has two amps each handling one
channel and we have one amp handling two channels.

Change-Id: Iece3472f45ee5330893e387f6b77af769a21e8b3