dlx: Remove NEW_LIBRIL_HTC
As far as I can tell, this is for different HTC RIL proprietary
libs.  When this was added it says it is for newer libhtc_ril.so.
We on't have that proprietary lib.

Furthermore, this option is not being used for the M7 phones
which indicates we likely shouldn't be using it either.

Change-Id: Id08699b316f94e3707ef34fbec7630b6f8be5189