Update patch set 1
Patch Set 1:

I suppose that is a very effective strategy...

This commit (and two others that I pushed yesterday) didn't seem to make it to github.  They say that they are merged, cmxlog shows them in last night's build but the build doesn't contain the changes and they aren't in the real github repo.

Any idea what could have happened there?  And, more importantly, how to fix it?

Patch-set: 1