dlx: Add performance profiles to recovery
When performing intensive work, bring all cores online and make
the first two cores run with the performance governor and the
second two with OnDemand.

This allows a .tar backup/restore to not leave two cores pegged
at max freq for no reason while still allowing all 4 to be used
for a .tar.gz backup.

Change-Id: I001c7f0194c623f5b3a06dc252c11d8e5ee2aa4c