Update patch set 2
Patch Set 2:

To answer my own question via xda:

@crpalmer
A lot of users having issues on the m7vzw with youtube videos not working right or data lag. Has any users reported you seen with this. I think it could be the stock RIL we are using.. I am on a test build running with the HTCQualcommRIL and everything is working fine again. I will submit a revert on the gerrit just for sakes if you end up with better data results. http://review.cyanogenmod.org/#/c/58273/

Patch-set: 2