Update patch set 1
Patch Set 1:

There's a difference of opinion on it. My opinion is that a device that has enough space for it should use it. This was the default on 10.2 as well.

Although I'll probably get chokeslammed for this, I'll just pick it for my builds. It annoys me to waste this space.

Patch-set: 1
Reviewer: Dan Pasanen <3671@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0