Update patch set 3
Patch Set 3:

Ah, thanks.

Hard to say about the revert, it seems we're taking a different approach to device-specific hacks making it into the main repos this time around. Which is good in a way I guess, device-specific hax should be in device trees. The bad part is that we as maintainers seem to have been a bit blindsided by the idea.

Patch-set: 3