Update patch set 1
Patch Set 1:

Seems to work as intended.

This is causing issues with selinux enforcing in lollipop. It seems when it tries to mkdir /vendor, shit starts dying. Since we can just symlink the firmware to /system/vendor/firmware/* and /system/vendor gets automatically symlinked to /vendor (see /system/core) we can just do it this way and not use the crappy script.

Patch-set: 1