Update patch set 13
Patch Set 13: Code-Review+1

I haven't tried running this, but I did read through it and I'm good with shipping this and dealing with any problems later rather than worrying about a whole lot of extra testing of it.

Two suggestions which could make them slightly better though.

Is there a better type name than cm3602?  It's a prox and light sensor, isn't it?  If so, that's a funny name and that's why I'm fine with shipping that as-is.

There are two rules matching on vfat.  I assume that is really firmware.  Any chance of labeling those partitions to make the rule clearer and more restrictive?  I tried to do so on mondrianwifi and failed though (I think I need file_context mount options to be able to do it, and those aren't working).

That is, I shipped mondrianwifi with vfat so either ship it or show me how to fix my other device!  :-)

Patch-set: 13